requirements.txt
asttokens==2.4.1
auto-click-auto==0.1.4
blinker==1.7.0
click==8.1.7
decorator==5.1.1
dnspython==2.5.0
email-validator==2.1.0.post1
executing==2.0.1
Flask==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
